pf-net/requirement.txt
numpy==1.16.1
paddlepaddle-gpu==1.8.3.post97
open3d==0.9.0
